rename docker_client to docker_client_maker, modify the relative test files
