+++ tests/unit/cli/docker_client_test.py
@@ -5,7 +5,7 @@ import os
 import mock
 from tests import unittest
 
-from fig.cli import docker_client 
+from fig.cli import docker_client_maker
 
 
 class DockerClientTestCase(unittest.TestCase):
@@ -13,4 +13,4 @@ class DockerClientTestCase(unittest.TestCase):
     def test_docker_client_no_home(self):
         with mock.patch.dict(os.environ):
             del os.environ['HOME']
-            docker_client.docker_client_maker()
+            docker_client_maker.docker_client_maker()
